float to double for NO BENEFIT

+++ Autopilot.cpp
@@ -9,9 +9,9 @@
 #include "controls.h"
 
 const char filename[] = "meta/C5A.xlsx";
-Eigen::Matrix<float, 4, 1> Controls;
-const float dt = 0.01;
-const float tfinal = 20;
+Eigen::Matrix<double, 4, 1> Controls;
+const double dt = 0.01;
+const double tfinal = 20;
 
 bool fileExists(const char* path) {
     struct stat buffer;
@@ -38,17 +38,17 @@ int main() {
     RBDsolve RBD(c5a, Controls);
 
     // Initial state vector: [v(3), omega(3), euler(3)]
-    Eigen::Matrix<float, 9, 1> initial_state;
+    Eigen::Matrix<double, 9, 1> initial_state;
     initial_state << c5a.V0(0), c5a.V0(1), c5a.V0(2),
                       c5a.omega0(0), c5a.omega0(1), c5a.omega0(2),
                       c5a.euler0(0), c5a.euler0(1), c5a.euler0(2);
     
     // Setup and run RK4 integration
     rk4 rk4Solver(dt, tfinal);
-    Eigen::Matrix<float, 9, 1>* results = rk4Solver.rk4_solver(RBD, initial_state);
+    Eigen::Matrix<double, 9, 1>* results = rk4Solver.rk4_solver(RBD, initial_state);
 
     int N_steps = (int)(tfinal / dt);
-    Eigen::Matrix<float, 9, 1> final_state = results[N_steps];
+    Eigen::Matrix<double, 9, 1> final_state = results[N_steps];
     
     std::cout << "\n=== Final State (t=" << tfinal << "s) ===" << std::endl;
     std::cout << "Velocity (v_x, v_y, v_z): "
